requirements.txt
numpy==1.26.4
pandas==2.1.4
matplotlib==3.7.1
seaborn==0.13.1
tensorflow==2.17.0
pillow>=9.4.0
mtcnn==0.1.1
openpyxl>=3.1.5
streamlit==1.38.0
opencv-python>=4.5.0
huggingface-hub>=0.31.1